+++ lib/domain/admin/all_orders/all_orders_cubit.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:moxy/constant/order_constants.dart';
+import 'package:moxy/data/repositories/filter_repository.dart';
 import 'package:moxy/data/repositories/order_repository.dart';
 import 'package:moxy/domain/admin/all_orders/all_orders_state.dart';
 import 'package:moxy/utils/common.dart';
@@ -10,7 +11,8 @@ import '../../../services/get_it.dart';
 import '../../mappers/order_mapper.dart';
 
 class AllOrdersCubit extends Cubit<AllOrdersState> {
-  final orderRepository = locate<OrderRepository>();
+  final filterRepository = locate<FilterRepository>();
+  final ordersRepository = locate<OrderRepository>();
   final orderMapper = locate<OrderMapper>();
 
   late final StreamSubscription _filterOrderSubscription;
@@ -22,7 +24,7 @@ class AllOrdersCubit extends Cubit<AllOrdersState> {
 
   void _subscribe() {
     _filterOrderSubscription =
-        orderRepository.filterParamsStream.listen((filterState) {
+        filterRepository.filterParamsStream.listen((filterState) {
       allOrders();
     });
   }
@@ -36,7 +38,7 @@ class AllOrdersCubit extends Cubit<AllOrdersState> {
   void allOrders() async {
     try {
       emit(state.copyWith(isLoading: true));
-      final result = await orderRepository.getAllOrders();
+      final result = await ordersRepository.getAllOrders();
       checkFilters();
       result.when((success) {
         emit(state.copyWith(allOrders: success, isLoading: false));
@@ -49,7 +51,7 @@ class AllOrdersCubit extends Cubit<AllOrdersState> {
   }
 
   checkFilters() {
-    final filters = orderRepository.getFilterParams();
+    final filters = filterRepository.getFilterParams();
     emit(state.copyWith(
       paymentFilter: filters.paymentType,
       deliveryFilter: filters.deliveryType,
@@ -59,22 +61,22 @@ class AllOrdersCubit extends Cubit<AllOrdersState> {
   }
 
   void clearDeliveryTypeFilter() async {
-    orderRepository.clearDeliveryTypeFilter();
+    filterRepository.clearDeliveryTypeFilter();
     emit(state.copyWith(deliveryFilter: FilterDeliveryType.empty));
   }
 
   void clearPaymentTypeFilter() {
-    orderRepository.clearPaymentTypeFilter();
+    filterRepository.clearPaymentTypeFilter();
     emit(state.copyWith(paymentFilter: FilterPaymentType.empty));
   }
 
   void clearStatusFilter() {
-    orderRepository.clearStatusFilter();
+    filterRepository.clearStatusFilter();
     emit(state.copyWith(statusFilter: ''));
   }
 
   void clearDateRangeFilter() {
-    orderRepository.clearDateRangeFilter();
+    filterRepository.clearDateRangeFilter();
     emit(state.copyWith(dateRangeFilter: null));
   }
 }
